#37 [refact] : uistate common모듈로 이동
