+++ core/common/src/main/java/com/teamwable/common/uistate/UiState.kt
@@ -1,17 +1,16 @@
-package com.teamwable.ui.util
-
+package com.teamwable.common.uistate
 
 sealed interface UiState<out T> {
-    object Empty : UiState<Nothing>
+    data object Empty : UiState<Nothing>
 
-    object Loading : UiState<Nothing>
+    data object Loading : UiState<Nothing>
 
     data class Success<T>(
         val data: T,
     ) : UiState<T>
 
     data class Failure(
-        val msg: String,
+        val errorMessage: String,
     ) : UiState<Nothing>
 
     fun getUiStateModel(): UiStateModel {
@@ -19,7 +18,7 @@ sealed interface UiState<out T> {
             this is Empty,
             this is Loading,
             this is Success,
-            this is Failure
+            this is Failure,
         )
     }
 }
@@ -28,5 +27,5 @@ data class UiStateModel(
     val isEmpty: Boolean = false,
     val isLoading: Boolean = true,
     val isSuccess: Boolean = false,
-    val isFailure: Boolean = false
+    val isFailure: Boolean = false,
 )
